# public/docs/openapi.yaml
openapi: 3.0.3
info:
  title: 'KH Shop API'
  description: 'API for Web Application Assignment Year 4'
  version: 1.0.0
servers:
  -
    url: 'http://127.0.0.1:8000'
paths:
...
  '/api/products/{id}':
    get:
      summary: 'show the specific product'
      operationId: showTheSpecificProduct
      description: ''
      parameters: []
      responses:
        200:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  status: 200
                  message: 'Get product successfully'
                  data:
                    id: 37
                    name: 'Protein Powder'
                    description: 'Whey protein for muscle recovery.'
                    rate: '4.8'
                    image: 'https://www.placehold.co/270x400'
                    imported_from: GNC
                    star: 0
                    category_id: 5
                    regular_price: '59.99'
                    sale_price: '49.99'
                    created_at: '2025-03-11T15:48:24.000000Z'
                    updated_at: '2025-03-11T15:48:24.000000Z'
                properties:
                  status:
                    type: integer
                    example: 200
                  message:
                    type: string
                    example: 'Get product successfully'
                  data:
                    type: object
                    properties:
                      id:
                        type: integer
                        example: 37
                      name:
                        type: string
                        example: 'Protein Powder'
                      description:
                        type: string
                        example: 'Whey protein for muscle recovery.'
                      rate:
                        type: string
                        example: '4.8'
                      image:
                        type: string
                        example: 'https://www.placehold.co/270x400'
                      imported_from:
                        type: string
                        example: GNC
                      star:
                        type: integer
                        example: 0
                      category_id:
                        type: integer
                        example: 5
                      regular_price:
                        type: string
                        example: '59.99'
                      sale_price:
                        type: string
                        example: '49.99'
                      created_at:
                        type: string
                        example: '2025-03-11T15:48:24.000000Z'
                      updated_at:
                        type: string
                        example: '2025-03-11T15:48:24.000000Z'
      tags:
        - Endpoints
      security: []
    put:
      summary: 'edit the specific product'
      operationId: editTheSpecificProduct
      description: ''
      parameters: []
      responses: {  }
      tags:
        - Endpoints
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  description: 'Must not be greater than 100 characters.'
                  example: hqf
                  nullable: false
                rate:
                  type: integer
                  description: 'Must be at least 0. Must not be greater than 10.'
                  example: 3
                  nullable: false
                image:
                  type: string
                  description: 'Must not be greater than 2048 characters.'
                  example: zml
                  nullable: true
                description:
                  type: string
                  description: 'Must not be greater than 255 characters.'
                  example: 'Autem neque nesciunt pariatur dolores corporis quis illo.'
                  nullable: false
                imported_from:
                  type: string
                  description: 'Must not be greater than 100 characters.'
                  example: bowuewncuwddhawin
                  nullable: false
                category_id:
                  type: integer
                  description: ''
                  example: 5
                  nullable: false
              required:
                - name
                - rate
                - description
                - imported_from
                - category_id
      security: []
    delete:
      summary: 'delete the specific product'
      operationId: deleteTheSpecificProduct
      description: ''
      parameters: []
      responses: {  }
      tags:
        - Endpoints
      security: []
    parameters:
      -
        in: path
        name: id
        description: 'The ID of the product.'
        example: 37
        required: true
        schema:
          type: integer
...